pyproject.toml
[project]
name = "prototype-analyzer"
version = "0.1.0"
description = "prototype-analyzer using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.140.0,<1.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0"
]

[project.scripts]
prototype-analyzer = "prototype_analyzer.main:run"
run_crew = "prototype_analyzer.main:run"
train = "prototype_analyzer.main:train"
replay = "prototype_analyzer.main:replay"
test = "prototype_analyzer.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
packages = ["src/prototype_analyzer"]

[tool.hatch.build.targets.wheel]
packages = ["src/prototype_analyzer"]

[tool.crewai]
type = "crew"
